0.x-split-integration-into-own-suite
