+++ .github/workflows/tests.yml
@@ -36,4 +36,4 @@ jobs:
         run: composer update --prefer-dist --no-interaction --no-progress --with="illuminate/contracts=^${{ matrix.laravel }}"
 
       - name: Execute tests
-        run: vendor/bin/phpunit --testsuite Unit
+        run: vendor/bin/phpunit --testsuite Unit,Feature
