+++ tests/Integration/AiProviderEnumIntegrationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Integration;
 
 use Illuminate\Support\Facades\Event;
 use Laravel\Ai\Audio;
