+++ tests/Integration/RerankingIntegrationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Integration;
 
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Event;
